Separate out panel 3 into its own file
